Remove windows.h from hkey_holder.hpp by inlining that into the .cpp
